Update the shelf and its operations
The most important change to be documented here is the removal of the ability to create a shelf instance entirely in memory. It is just overengineering at its finest. It also adds a lot more checks that is rarely needed in practical application anyway.

There are also more atomic operations derived from the main operations to further reduce the code.