Update the subject, note, and the shelf data operations
The most notable change is the shelf now uses a database pool instead of a single connection for multithreading. It uses the R2D2 dependency crate for now but a dedicated function that creates a connection would be better to reduce the dependent code for a bit.

The second most notable change is the addition of more atomic operations for the subject. I added those in for convenience. I also tried to make the subject to loosely couple with the shelf.