Update the notes and the shelf
The biggest notable change is the builder for the shelf which enables complex and dynamic configuration. The shelf constructor is also changed because of this.

The second big refactor is the addition of the export options since the parameters are getting numerous and might expand in the future. It also enables simple passing by reference for an array of shelf items.

Another update for the subject making the metadata file completely optional. Both subjects and notes now have a helper function for sorting out by date provided that they exist in the filesystem.