Revise the data structure of the profile
The profile now has a builder to enable complex and dynamic configuration similar to the shelf. Similarly, the shelf constructor has been updated. The export function now solely holds (or at least should solely hold) the responsibility of creating the profile in the filesystem.

P.S. The 'from' methods might use a little rework implemented as trait methods with the 'From' trait.