Refactor the codebase
Notable changes include the profile and the shelf are now independent of each other. Because of this change, there is also a change of workflow when using the program. The profile is more or less a configuration folder now and the shelf is more or less the current folder with a different terminology. This workflow now resembles more as described from the inspiration post at https://castel.dev/post/lecture-notes-3/

Another major change is the shelf being free of the database. As a consequence, the global ID system of shelf items has been removed.

Lastly, the major change is the usage of TOML instead of JSON as the config format. It is simply a change for the convenience and easier writing of the config files.

There is also a thread pool module but it is to be used in the future codebase. It is just committed for the sake of saving it.

The codebase is also more modularized for easier maintenance.