Implement the note filter for subjects (#2)
Implement the note filter system. 
The subject struct is also reduced to a single path. This also means the metadata is retrieved as its own data type.

An additional option in the executable named 'files' mainly used for filtering the notes of a subjects.

These updates has been documented in the manual.