Refactor codebase (#4)
Some components of the library is also moved to the binary crate instead like the compilation env and its related traits. The parallel execution is also now handled with the Rayon library instead of implementing its own thread pool.

The string case conversion now relies on the 'heck' library instead of implementing its own. It has the use cases I'm looking for and it has a small codebase so it is easy to derive and extend it if I want to create my own version.

Mainly, the template helpers are added. The command conversion is also added as a trait instead of a static function.